Cafeクラスを別のファイルに移動した｡
