+++ Cafe.py
@@ -1,0 +1,6 @@
+class Cafe(object):
+    def __init__(self, name):
+        self.name = name
+
+    def get_cafe_info(self):
+        return f"Name:{self.name}"
